refactor: separat error dialog stories
