+++ packages/admin-stories/src/docs/comet/components/ErrorHandling/ErrorDialog/ErrorDialog.stories.mdx
@@ -53,11 +53,11 @@ const ApolloProvider: React.FunctionComponent = ({ children }) => {
 Error Dialog can be used to display an error in a dialog. Try it out by simple pressing the Button below. You can also use the errorDialog with the `useErrorDialog` Hook
 
 <Canvas>
-    <Story id="stories-components-errorhandling-errordialog--manual-error-dialog" />
+    <Story id="stories-components-error-handling-error-dialog-manual-error-dialog--manual-error-dialog" />
 </Canvas>
 
 ### Automatic Error Dialog on graphQL Error
 
 <Canvas>
-    <Story id="stories-components-errorhandling-errordialog--automatic-error-on-graphql-error" />
+    <Story id="stories-components-error-handling-error-dialog-automatic-graphql-error-dialog--automatic-error-on-graphql-error" />
 </Canvas>
