+++ packages/admin-stories/src/docs/comet/components/ErrorHandling/ErrorDialog/stories/AutomaticGraphqlErrorDialog.stories.tsx
@@ -1,39 +1,15 @@
 import { useQuery } from "@apollo/client";
-import { ErrorDialogContext, ErrorDialogContextProps } from "@comet/admin";
 import { Button, Typography } from "@material-ui/core";
 import { storiesOf } from "@storybook/react";
 import gql from "graphql-tag";
 import * as React from "react";
 
-import { apolloSwapiStoryDecorator } from "../../../../../apollo-swapi-story.decorator";
-import { errorDialogStoryProviderDecorator } from "../../../../../error-dialog-provider.decorator";
+import { apolloSwapiStoryDecorator } from "../../../../../../apollo-swapi-story.decorator";
+import { errorDialogStoryProviderDecorator } from "../error-dialog-provider.decorator";
 
-storiesOf("stories/components/errorhandling/errordialog", module)
+storiesOf("stories/components/Error Handling/Error Dialog/Automatic Graphql Error Dialog", module)
     .addDecorator(apolloSwapiStoryDecorator())
     .addDecorator(errorDialogStoryProviderDecorator())
-    .add("Manual Error Dialog", () => {
-        return (
-            <ErrorDialogContext.Consumer>
-                {(errorDialog: ErrorDialogContextProps | undefined) => {
-                    return (
-                        <Button
-                            variant={"contained"}
-                            color={"primary"}
-                            onClick={() => {
-                                errorDialog?.showError({
-                                    title: "Error",
-                                    userMessage: "You can close the error dialog by pressing ok",
-                                    error: "This is an error detail information e.g. stack trace, which is only shown in development mode",
-                                });
-                            }}
-                        >
-                            Show Error Dialog
-                        </Button>
-                    );
-                }}
-            </ErrorDialogContext.Consumer>
-        );
-    })
     .add("Automatic Error on Graphql Error", () => {
         const Story = () => {
             const [brokenQuery, setBrokenQuery] = React.useState(false);
